Added new files for the paged quest view
Initial code for each file. It's not functional yet
